splitting up some tests in prep for unires tests
