+++ int/unires/resolver_config.yml
@@ -1,0 +1,35 @@
+methods:
+  - "sov"
+  - "abt"
+  - "btcr"
+  - "erc725"
+  - "dom"
+  - "stack"
+  - "ethr"
+  - "web"
+  - "v1"
+  - "key"
+  - "ipid"
+  - "jolo"
+  - "hacera"
+  - "elem"
+  - "seraphid"
+  - "github"
+  - "ccp"
+  - "work"
+  - "ont"
+  - "kilt"
+  - "evan"
+  - "echo"
+  - "factom"
+  - "dock"
+  - "trust"
+  - "io"
+  - "bba"
+  - "bid"
+  - "schema"
+  - "ion"
+  - "ace"
+  - "gatc"
+  - "unisot"
+  - "icon"
